shift around a bit and finish up timeline
