+++ src/components/transactions/TransactionTimeline.tsx
@@ -1,69 +1,20 @@
+import format from 'date-fns/format';
+import { Pencil, Trash2 } from 'lucide-react';
 import React from 'react';
+import { Group } from 'react-aria-components';
 import { NavigateFunction } from 'react-router-dom';
 import type { Dispatch } from 'redux';
 import type { AnyAction } from 'starfx';
 import { useSelector } from 'starfx/react';
+import { tv } from 'tailwind-variants';
 
 import { transactionsInTimeline } from '~/src/store/selectors/transactions';
+import { transactionRemove } from '~/src/store/thunks';
+import { toHumanCurrency } from '~/src/store/utils/dineroUtils';
 
-import { TransactionFilter } from './utils';
+import { Button } from '~/src/elements/Button';
 
-const timeline = [
-  {
-    id: 1,
-    content: 'Applied to',
-    target: 'Front End Developer',
-    href: '#',
-    date: 'Sep 20',
-    datetime: '2020-09-20',
-    icon: 'UserIcon',
-    iconBackground: 'bg-gray-400'
-  },
-  {
-    id: 2,
-    content: 'Advanced to phone screening by',
-    target: 'Bethany Blake',
-    href: '#',
-    date: 'Sep 22',
-    datetime: '2020-09-22',
-    icon: 'HandThumbUpIcon',
-    iconBackground: 'bg-blue-500'
-  },
-  {
-    id: 3,
-    content: 'Completed phone screening with',
-    target: 'Martha Gardner',
-    href: '#',
-    date: 'Sep 28',
-    datetime: '2020-09-28',
-    icon: 'CheckIcon',
-    iconBackground: 'bg-green-500'
-  },
-  {
-    id: 4,
-    content: 'Advanced to interview by',
-    target: 'Bethany Blake',
-    href: '#',
-    date: 'Sep 30',
-    datetime: '2020-09-30',
-    icon: 'HandThumbUpIcon',
-    iconBackground: 'bg-blue-500'
-  },
-  {
-    id: 5,
-    content: 'Completed interview with',
-    target: 'Katherine Snyder',
-    href: '#',
-    date: 'Oct 4',
-    datetime: '2020-10-04',
-    icon: 'CheckIcon',
-    iconBackground: 'bg-green-500'
-  }
-];
-
-function classNames(...classes) {
-  return classes.filter(Boolean).join(' ');
-}
+import { navigateToTransactionForm, TransactionFilter } from './utils';
 
 export const TransactionTimeline = ({
   navigate,
@@ -75,64 +26,109 @@ export const TransactionTimeline = ({
   transactionFilter: TransactionFilter;
 }) => {
   const transactionsTimeline = useSelector(transactionsInTimeline);
-  {
-    /* <div>
-          {transactionsTimeline.map((d) => (
-            <div key={d.date.toISOString()}>
-              {d.date.toISOString()}{' '}
-              {d.transactions.map((t) => t.description).join(',')}
-            </div>
-          ))}
-        </div> */
-  }
 
   return (
     <div className="flow-root">
-      <ul role="list" className="-mb-8">
-        {timeline.map((event, eventIdx) => (
-          <li key={event.id}>
-            <div className="relative pb-8">
-              {eventIdx !== timeline.length - 1 ? (
-                <span
-                  aria-hidden="true"
-                  className="absolute left-4 top-4 -ml-px h-full w-0.5 bg-gray-200"
-                />
-              ) : null}
-              <div className="relative flex space-x-3">
-                <div>
-                  <span
-                    className={classNames(
-                      event.iconBackground,
-                      'flex size-8 items-center justify-center rounded-full ring-8 ring-white'
-                    )}
-                  >
-                    <event.icon
-                      aria-hidden="true"
-                      className="size-5 text-white"
-                    />
-                  </span>
+      <ul role="list" className="space-y-6">
+        {transactionsTimeline
+          .filter((d) => d.transactions.length > 0)
+          .map((d, eventIdx) => (
+            <li key={eventIdx} className="relative flex gap-x-4">
+              {eventIdx === transactionsTimeline.length - 1 ? (
+                <div className="h-6 absolute left-0 top-0 flex w-6 justify-center">
+                  <div className="w-px bg-gray-200" />
                 </div>
-                <div className="flex min-w-0 flex-1 justify-between space-x-4 pt-1.5">
-                  <div>
-                    <p className="text-sm text-gray-500">
-                      {event.content}{' '}
-                      <a
-                        href={event.href}
-                        className="font-medium text-gray-900"
-                      >
-                        {event.target}
-                      </a>
-                    </p>
-                  </div>
-                  <div className="whitespace-nowrap text-right text-sm text-gray-500">
-                    <time dateTime={event.datetime}>{event.date}</time>
-                  </div>
+              ) : (
+                <div className="-bottom-6 absolute left-0 top-0 flex w-6 justify-center">
+                  <div className="w-px bg-gray-200" />
                 </div>
-              </div>
-            </div>
-          </li>
-        ))}
+              )}
+              <>
+                <div className="relative flex size-6 flex-none items-center justify-center bg-white">
+                  <div className="size-1.5 rounded-full bg-gray-100 ring-1 ring-gray-300" />
+                </div>
+                <div className="flex-auto py-0.5 text-xs/5 text-gray-500">
+                  <span className="font-medium text-gray-900">
+                    {d.date.toLocaleDateString()}
+                  </span>{' '}
+                  {format(d.date, 'EEEE')}
+                  {d.transactions
+                    .filter(
+                      (transaction) =>
+                        transactionFilter === 'all' ||
+                        transactionFilter.has(transaction.type)
+                    )
+                    .map((t) => (
+                      <p className="flex gap-3 max-w-md text-gray-700">
+                        <span className="text-black flex-none">
+                          {t.raccount}
+                        </span>
+                        <span className="flex-none text-gray-800">
+                          {t.category}
+                        </span>
+                        <span className="flex-auto">
+                          {t.description ? `${t.description}` : ''}
+                        </span>
+                        <span
+                          className={transactionTypeColor({ type: t.type })}
+                        >
+                          {transactionTypeToSymbol(t.type)}
+                          {toHumanCurrency(t.value)}
+                        </span>
+                        <Group aria-label="Actions" className="space-x-1">
+                          <Button
+                            aria-label="Modify"
+                            className="py-0.5 px-0.5"
+                            onPress={() =>
+                              navigate(
+                                '/transactions/set',
+                                navigateToTransactionForm(t)
+                              )
+                            }
+                          >
+                            <Pencil className="max-h-3" />
+                          </Button>
+                          <Button
+                            aria-label="Delete"
+                            className="py-0.5 px-0.5"
+                            onPress={() =>
+                              dispatch(transactionRemove({ id: t.id }))
+                            }
+                          >
+                            <Trash2 className="max-h-3" />
+                          </Button>
+                        </Group>
+                      </p>
+                    ))}
+                </div>
+              </>
+            </li>
+          ))}
       </ul>
     </div>
   );
+};
+
+const transactionTypeColor = tv({
+  base: 'flex-none',
+  variants: {
+    type: {
+      income: 'text-green-700',
+      transfer: 'text-purple-700',
+      expense: 'text-red-700'
+    }
+  }
+});
+
+const transactionTypeToSymbol = (type: string) => {
+  switch (type) {
+    case 'income':
+      return '+';
+    case 'expense':
+      return '-';
+    case 'transfer':
+      return '⇄';
+    default:
+      return '';
+  }
 };
