+++ src/store/selectors/chartData.ts
@@ -1,51 +1,8 @@
-import { parseISO, isSameDay, isWithinInterval, addDays } from 'date-fns';
-import eachDayOfInterval from 'date-fns/fp/eachDayOfInterval/index.js';
-import { toDecimal, type Dinero } from 'dinero.js';
+import { toDecimal } from 'dinero.js';
 import { createSelector } from 'starfx';
 
-import { schema, Transaction } from '../schema';
-import { nextTransaction } from '../thunks/transactionReoccurrence';
-
-export const eachDay = createSelector(schema.chartRange.select, (chartRange) =>
-  eachDayOfInterval(chartRange)
-);
-
-export interface TransactionWithSeed extends Transaction {
-  seedDate: Date;
-  occurredInSeed: number;
-}
-export const transactionsWithSeed = createSelector(
-  schema.chartRange.select,
-  schema.transactions.selectTableAsList,
-  (chartRange, transactions) =>
-    transactions.map((transaction) => {
-      const nextTransactionFn = nextTransaction(transaction.rtype);
-      const { date, occurred: occurredInSeed } = findSeed({
-        transaction,
-        y: transaction.value,
-        // back off one day to start outside the interval
-        date: parseISO(transaction.start),
-        nextTransactionFn,
-        interval: chartRange,
-        occurred: 0
-      });
-      return { ...transaction, seedDate: date, occurredInSeed };
-    })
-);
-
-export const chartableData = createSelector(
-  eachDay,
-  transactionsWithSeed,
-  (allDates, transactions) => {
-    return transactions.map((transaction) => {
-      const { data, allTransactionEvents } = resolveBarChartData({
-        transaction,
-        allDates
-      });
-      return { transaction, data, allTransactionEvents };
-    });
-  }
-);
+import { Transaction } from '../schema';
+import { chartableData } from './transactions';
 
 export const barChartTransactions = createSelector(
   chartableData,
@@ -59,7 +16,6 @@ export const barChartTransactions = createSelector(
     );
 
     const incomeStacked = stackTransactions(income);
-    console.log({ incomeStacked });
     const expensesStacked = stackTransactions(expenses);
     const transfersStacked = stackTransactions(transfers);
     const maxValue = Math.max(
@@ -118,89 +74,4 @@ const stackTransactions = (
     return { ...item, stacked };
   });
   return { maxValue, stack: transactionStack };
-};
-
-export function resolveBarChartData({
-  allDates,
-  transaction
-}: {
-  allDates: Date[];
-  transaction: any;
-}) {
-  const nextTransactionFn = nextTransaction(transaction.rtype);
-
-  const allTransactionEvents = [] as Date[];
-  const next = {
-    transaction,
-    occurrences: transaction.occurrences,
-    date: transaction.seedDate,
-    nextY: transaction.value
-  };
-  let occurred = transaction.occurredInSeed - 1; // we will capture once instance on first loop
-  const stack = allDates.map((day) => {
-    let y = null;
-    if (
-      isSameDay(day, next.date) &&
-      (transaction.occurrences === 0 || occurred < transaction.occurrences)
-    ) {
-      allTransactionEvents.push(day);
-      y = next.nextY;
-
-      const { date, y: calculatedY } = nextTransactionFn({
-        ...next,
-        seedDate: addDays(next.date, 1)
-      });
-      if (isSameDay(date, next.date))
-        throw new Error(
-          'same date, recursive calc, we should not hit this error'
-        );
-      // save data for next value
-      next.nextY = calculatedY;
-      next.date = date;
-      occurred += 1;
-    }
-    return { date: day, y };
-  });
-
-  return { data: stack, allTransactionEvents };
-}
-
-export const findSeed = ({
-  nextTransactionFn,
-  transaction,
-  date,
-  y,
-  interval,
-  occurred
-}: {
-  nextTransactionFn: any;
-  transaction: any;
-  date: Date;
-  y: Dinero<number>;
-  interval: Interval;
-  occurred: number;
-}): { date: Date; nextY: Dinero<number>; occurred: number } => {
-  // a transaction function has to run and mark an occurenace to have found the seed date
-  //  so we don't blindly use the transaction start date as the seed except for daily as
-  //  the start date dictates the start of a cycle
-  if (
-    isWithinInterval(date, interval) &&
-    (occurred !== 0 || transaction.rtype === 'day')
-  )
-    return { date, nextY: y, occurred };
-  const { date: nextDate, y: nextY } = nextTransactionFn({
-    transaction,
-    seedDate: date,
-    occurrences: transaction.occurrences
-  });
-
-  return findSeed({
-    nextTransactionFn,
-    transaction,
-    interval,
-    y: nextY,
-    // to avoid getting stuck generating the same day over and over
-    date: isSameDay(date, nextDate) ? addDays(nextDate, 1) : nextDate,
-    occurred: occurred + 1
-  });
 };
