+++ src/store/selectors/transactions.ts
@@ -1,17 +1,45 @@
+import { parseISO } from 'date-fns';
+import eachDayOfInterval from 'date-fns/fp/eachDayOfInterval/index.js';
 import { createSelector } from 'starfx';
 
-import { schema, type Account } from '~/src/store/schema.ts';
+import { schema, Transaction, type Account } from '~/src/store/schema.ts';
 
+import { nextTransaction } from '../thunks/transactionReoccurrence';
 import {
-  chartableData,
-  eachDay,
-  transactionsWithSeed,
-  TransactionWithSeed
-} from './chartData';
+  extrapolateTransactionOccurrences,
+  findSeed
+} from '../utils/extrapolateDates';
+
+export const eachDay = createSelector(schema.chartRange.select, (chartRange) =>
+  eachDayOfInterval(chartRange)
+);
+
+export interface TransactionWithSeed extends Transaction {
+  seedDate: Date;
+  occurredInSeed: number;
+}
+export const transactionsWithSeed = createSelector(
+  schema.chartRange.select,
+  schema.transactions.selectTableAsList,
+  (chartRange, transactions) =>
+    transactions.map((transaction) => {
+      const nextTransactionFn = nextTransaction(transaction.rtype);
+      const { date, occurred: occurredInSeed } = findSeed({
+        transaction,
+        y: transaction.value,
+        // back off one day to start outside the interval
+        date: parseISO(transaction.start),
+        nextTransactionFn,
+        interval: chartRange,
+        occurred: 0
+      });
+      return { ...transaction, seedDate: date, occurredInSeed };
+    })
+);
 
 export interface TransactionWithAccount extends TransactionWithSeed {
   raccountMeta: Account;
-  transferInMeta: Account;
+  transferInMeta?: Account;
 }
 
 export const transactionsWithAccounts = createSelector(
@@ -40,12 +68,26 @@ export const transactionsWithAccounts = createSelector(
   }
 );
 
+export const chartableData = createSelector(
+  eachDay,
+  transactionsWithAccounts,
+  (allDates, transactions) => {
+    return transactions.map((transaction) => {
+      const { data, allTransactionEvents } = extrapolateTransactionOccurrences({
+        transaction,
+        allDates
+      });
+      return { transaction, data, allTransactionEvents };
+    });
+  }
+);
+
 export const transactionsInTimeline = createSelector(
   eachDay,
   chartableData,
   (allDates, transactionsWithStacks) => {
     const datesWithTransactions = allDates.map((date) => {
-      const thisDay = { date, transactions: [] as TransactionWithSeed[] };
+      const thisDay = { date, transactions: [] as TransactionWithAccount[] };
       for (const transaction of transactionsWithStacks) {
         if (transaction.allTransactionEvents.includes(date)) {
           thisDay.transactions.push(transaction.transaction);
